finance-tracker-app-main/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pandas
openpyxl
python-docx
PyPDF2
reportlab
